Add files via upload
Added the main feature of being able to summarize text, and added a display box for the results. The current shortcut to trigger this is ctrl+alt+z. Note that to test this I was simply putting my apikey in as a variable, this has been removed in this version because of privacy so to run this version a key must be entered manually into the code.